Migrate jpa-idempotent-repository example from deprecated derby container image to MariaDB
