+++ jpa-idempotent-repository/src/main/resources/application.properties
@@ -26,7 +26,7 @@ timer.repeatCount = 0
 quarkus.http.port = 8085
 quarkus.http.test-port = 8085
 
-derby-hostname=localhost
+db-hostname=localhost
 
 # Uncomment if your application image is to be pushed to an external registry
 #quarkus.container-image.registry=my.docker-registry.net
@@ -37,7 +37,14 @@ derby-hostname=localhost
 #quarkus.kubernetes-client.trust-certs=true
 
 quarkus.kubernetes.image-pull-policy=IfNotPresent
-quarkus.kubernetes.env.vars.derby-hostname=derby-database
+quarkus.kubernetes.env.vars.db-hostname=mariadb-database
+
+# Enable the application to resolve the MariaDB credentials via a secret
+quarkus.kubernetes.env.secrets=mariadb-secret
+quarkus.kubernetes.env.mapping.db-user.from-secret=mariadb-secret
+quarkus.kubernetes.env.mapping.db-user.with-key=db-user
+quarkus.kubernetes.env.mapping.db-password.from-secret=mariadb-secret
+quarkus.kubernetes.env.mapping.db-password.with-key=db-password
 
 # Uncomment to set resource limits
 #quarkus.kubernetes.resources.requests.memory=64Mi
@@ -47,7 +54,7 @@ quarkus.kubernetes.env.vars.derby-hostname=derby-database
 
 # OpenShift
 quarkus.openshift.image-pull-policy=IfNotPresent
-quarkus.openshift.env.vars.derby-hostname=derby-database
+quarkus.openshift.env.vars.db-hostname=mariadb-database
 
 # Uncomment to set resource limits
 #quarkus.openshift.resources.requests.memory=64Mi
@@ -56,12 +63,13 @@ quarkus.openshift.env.vars.derby-hostname=derby-database
 #quarkus.openshift.resources.limits.cpu=1000m
 
 # JPA
-quarkus.datasource.db-kind=derby
-%dev.quarkus.datasource.db-kind=h2
-
-quarkus.datasource.jdbc.url=jdbc:derby://${derby-hostname}:1527/my-db
-%dev.quarkus.datasource.jdbc.url=jdbc:h2:mem:my-db;DB_CLOSE_DELAY=-1
-
+quarkus.datasource.db-kind=mariadb
+quarkus.datasource.devservices.db-name=my-db
 quarkus.datasource.jdbc.max-size=8
+%prod.quarkus.datasource.jdbc.url=jdbc:mariadb://${db-hostname}:3306/my-db
+%prod.quarkus.datasource.username=${db-user:mariadb}
+%prod.quarkus.datasource.password=${db-password:mariadb}
 
-quarkus.hibernate-orm.database.generation=drop-and-create
+# Flyway
+quarkus.flyway.migrate-at-start=true
+quarkus.flyway.schemas=my-db
