+++ jpa-idempotent-repository/src/main/resources/db/migration/V1.0.0__add_camel_message_processed.sql
@@ -15,6 +15,14 @@
 -- limitations under the License.
 --
 
-CONNECT 'jdbc:derby:my-db;create=true';
-CREATE TABLE CAMEL_MESSAGEPROCESSED ( processorName VARCHAR(255), messageId VARCHAR(100), createdAt TIMESTAMP, PRIMARY KEY (processorName, messageId) );
-CREATE SEQUENCE CAMEL_MESSAGEPROCESSED_SEQ AS INT MAXVALUE 999999 CYCLE;
+CREATE TABLE CAMEL_MESSAGEPROCESSED (
+    id BIGINT NOT NULL,
+    createdAt TIMESTAMP(6),
+    messageId VARCHAR(255),
+    processorName VARCHAR(255),
+    PRIMARY KEY (id)
+);
+
+CREATE SEQUENCE CAMEL_MESSAGEPROCESSED_SEQ START WITH 1 INCREMENT BY 50;
+
+ALTER TABLE CAMEL_MESSAGEPROCESSED ADD CONSTRAINT message_processed_key_constraint UNIQUE (processorName, messageId);
